refactor: move helper functions to backend
