+++ .gitignore
@@ -4,7 +4,7 @@ __pycache__/
 *$py.class
 
 # Samples
-# samples/*
+samples/*
 
 # VS Code
 .vscode
