+++ backend/document_storage.py
@@ -67,6 +67,3 @@ class DocumentStorage:
         return output.getvalue()
 
 
-def save_bytes_to_file(image_bytes: bytes, output_path: str):
-    with open(output_path, 'wb') as output_file:
-        output_file.write(image_bytes)
